api/requirements.txt
fastapi
uvicorn
qdrant-client
google-generativeai
pydantic
python-dotenv
